Create the getComputerChoice function.
The function picks a word between rock,
paper and scissors randomly. When the
Math.random() method returns a floating
point (decimal) greater than 0 but less
than 1, the number is multiplied by 3
(because there a 3 words and we seek 3
options). After being multiplied by 3,
it is rounded down with the Math.floor
Method and 1 is added to it.
The 1 is added to actually make it possible
to get 3. If Math.random returns 0.99,
multiplying it by 3 will result to 2.97.
Rounding it down will result to 2 and adding
1 will finally result to 3.

After that the conditioner statement returns
each of the words if the variable is equal
to 1, 2 or 3.